+++ PuzzleTokens.sketchplugin/Contents/Sketch/classes/DSExporter.js
@@ -26,6 +26,25 @@ const bordedLineEndMap2 = {
     [Style.LineEnd.Projecting]: "projecting"
 }
 
+const BLENDING_MODE_SKETCH_TO_CSS = {
+    [Style.BlendingMode.Normal]: "normal",
+    [Style.BlendingMode.Darken]: "darken",
+    [Style.BlendingMode.Multiply]: "multiply",
+    [Style.BlendingMode.ColorBurn]: "color-burn",
+    [Style.BlendingMode.Lighten]: "lighten",
+    [Style.BlendingMode.Screen]: "screen",
+    [Style.BlendingMode.ColorDodge]: "color-dodge",
+    [Style.BlendingMode.Overlay]: "overlay",
+    //"darken", [Style.BlendingMode.SoftLight]:  // Not supported in CSS
+    //"darken", [Style.BlendingMode.HardLight]:  // Not supported in CSS
+    [Style.BlendingMode.Difference]: "difference",
+    [Style.BlendingMode.Exclusion]: "exclusion",
+    [Style.BlendingMode.Hue]: "hue",
+    [Style.BlendingMode.Saturation]: "saturation",
+    [Style.BlendingMode.Color]: "color",
+    [Style.BlendingMode.Luminosity]: "luminosity"
+}
+
 
 const eol = ";\n"
 const pxeol = "px" + eol
@@ -56,7 +75,7 @@ class DSExporter {
         this.textStyles = {}
         this.runningForTokens = true
 
-        this.docName = this._clearCloudName(this.nDoc.cloudName())
+        this.fileName = "puzzle-tokens" //this._clearCloudName(this.nDoc.cloudName())
         this.pathTo = undefined
         this.format = undefined
         this.confOpts = undefined
@@ -80,8 +99,12 @@ class DSExporter {
     init() {
         this.pathTo = Settings.settingForKey(SettingKeys.PLUGIN_EXPORT_PATH_TO)
         if (undefined == this.pathTo) this.pathTo = ""
+
         this.format = Settings.settingForKey(SettingKeys.PLUGIN_EXPORT_FORMAT)
         if (undefined == this.format) this.format = Constants.EXPORT_FORMAT_LESS
+
+        this.openFinder = Settings.settingForKey(SettingKeys.PLUGIN_EXPORT_OPEN_FINDER)
+        if (null == this.openFinder) this.openFinder = true
 
         this.confExportLibStyles = Settings.settingForKey(SettingKeys.PLUGIN_EXPORT_LIB_STYLES) == 1
 
@@ -91,6 +114,8 @@ class DSExporter {
         if (null == this.confOpts.fontSizeTokens) this.confOpts.fontSizeTokens = true
         if (null == this.confOpts.fontWeightTokens) this.confOpts.fontWeightTokens = true
         if (null == this.confOpts.fontFamilyTokens) this.confOpts.fontFamilyTokens = true
+
+
 
         // init global variable
         app = this
@@ -140,6 +165,7 @@ class DSExporter {
             if (success) {
                 track(TRACK_EXPORT_COMPLETED)
                 this.UI.message("Completed")
+                if (this.openFinder) this._openFinder()
             }
         }
 
@@ -147,6 +173,10 @@ class DSExporter {
     }
 
     // Internal
+
+    _openFinder() {
+        NSWorkspace.sharedWorkspace().openFile(this.pathTo);
+    }
 
     _clearCloudName(cloudName) {
         let name = cloudName
@@ -177,7 +207,7 @@ class DSExporter {
 
     _showDialog() {
         const dialog = new UIDialog("Export Styles", NSMakeRect(0, 0, 800, 300), "Export",
-            "Export all text styles to a text file")
+            "Export all text styles to \"" + this.fileName + "\".[less|scss] file")
         //dialog.removeLeftColumn()
         dialog.leftColWidth = 200
 
@@ -190,6 +220,8 @@ class DSExporter {
 
         dialog.addLeftLabel("", "File Format")
         dialog.addRadioButtons("format", "", this.format, ["LESS", "SCSS"], 250)
+
+        dialog.addCheckbox("openFinder", "Open Finder window on completion", this.openFinder)
 
         dialog.addDivider()
 
@@ -223,6 +255,7 @@ class DSExporter {
             this.confExportLibStyles = dialog.views['exportLibStyles'].state() == 1
             this.less = this.format == Constants.EXPORT_FORMAT_LESS
             this.scss = this.format == Constants.EXPORT_FORMAT_SCSS
+            this.openFinder = dialog.views['openFinder'].state() == 1
             this.confOpts.colorTokens = dialog.views['colorTokens'].state() == 1
             this.confOpts.fontSizeTokens = dialog.views['fontSizeTokens'].state() == 1
             this.confOpts.fontWeightTokens = dialog.views['fontWeightTokens'].state() == 1
@@ -232,6 +265,7 @@ class DSExporter {
 
             Settings.setSettingForKey(SettingKeys.PLUGIN_EXPORT_PATH_TO, this.pathTo)
             Settings.setSettingForKey(SettingKeys.PLUGIN_EXPORT_FORMAT, this.format)
+            Settings.setSettingForKey(SettingKeys.PLUGIN_EXPORT_OPEN_FINDER, this.openFinder)
             Settings.setSettingForKey(SettingKeys.PLUGIN_EXPORT_OPTS, this.confOpts)
             Settings.setSettingForKey(SettingKeys.PLUGIN_EXPORT_LIB_STYLES, this.confExportLibStyles)
             break
@@ -248,7 +282,7 @@ class DSExporter {
 
         let res = tokensText + textStylesText
 
-        const fullPathTo = this.pathTo + "/" + this.docName + this.def.ext
+        const fullPathTo = this.pathTo + "/" + this.fileName + this.def.ext
         Utils.writeToFile(res, fullPathTo);
 
         return true
@@ -548,6 +582,16 @@ class DSExporter {
         } else {
             return ""
         }
+        //
+        if (Style.BlendingMode.Normal != sStyle.blendingMode) {
+            const cssValue = BLENDING_MODE_SKETCH_TO_CSS[sStyle.blendingMode]
+            if (undefined == cssValue) {
+                this.logError('Can not set CSS mix-blend-mode: for Sketch blendingMode ' + sStyle.blendingMode)
+            } else {
+                res += spaces + "mix-blend-mode: " + cssValue + eol
+            }
+        }
+        //
         return res
     }